Require all the things as needed in tests
create resque/globals to avoid having to require main resque module
(this is probably a sign that the remaining things in the resque module
are un-tested)
