+++ lib/resque/failure/redis.rb
@@ -1,5 +1,8 @@
 require 'time'
+require 'resque/failure'
+require 'resque/failure/base'
 require 'resque/failure/each'
+require 'resque/job'
 
 module Resque
   module Failure
