+++ test/resque/core_ext/string_test.rb
@@ -1,5 +1,7 @@
 require 'test_helper'
 
+require 'resque/core_ext/string'
+
 describe String do
   it "#constantize" do
     assert_same Kernel, "Kernel".constantize
