+++ test/resque/failure/each_test.rb
@@ -1,7 +1,14 @@
 require 'test_helper'
+
+require 'resque/globals'
 require 'resque/failure/redis'
 
 describe Resque::Failure::Each do
+  before do
+    #Previously this was getting set by the CLI test being run first
+    #Should the knowledge of the default 'localhost:6379/resque' be consolidated somewhere?
+    Resque.redis = "localhost:6379/resque"
+  end
   after do
     Resque::Failure.clear
   end
