+++ test/resque/job_performer_test.rb
@@ -1,3 +1,4 @@
+require 'resque/job_performer'
 require 'test_helper'
 
 describe Resque::JobPerformer do
